refactor(backend): split mcp service into mcp package
Move MCP logic into services/mcp/ (exceptions, templates, config, agents, endpoint operations, workflows).
Retarget unit test mock paths.
