--- apps/backend/src/rhesis/backend/app/services/mcp_service.py
@@ -1,675 +1,0 @@
-"""MCP service for generic integration using MCPAgent."""
-
-import json
-import logging
-import os
-import uuid
-from datetime import datetime
-from pathlib import Path
-from typing import Any, Dict, List, Optional
-
-import jinja2
-from fastapi import HTTPException
-from sqlalchemy.orm import Session
-
-from rhesis.backend.app import crud, schemas
-from rhesis.backend.app.constants import DEFAULT_GENERATION_MODEL
-from rhesis.backend.app.models.user import User
-from rhesis.backend.app.utils.database_exceptions import ItemDeletedException
-from rhesis.backend.app.utils.observability import get_test_context
-from rhesis.sdk.decorators import endpoint
-from rhesis.sdk.services.mcp import MCPAgent, MCPClientFactory
-from rhesis.sdk.services.mcp.exceptions import (
-    MCPApplicationError,
-    MCPConfigurationError,
-    MCPError,
-)
-from rhesis.sdk.services.mcp.observable_agent import ObservableMCPAgent
-
-logger = logging.getLogger(__name__)
-
-# Initialize Jinja2 environment for loading prompt templates
-TEMPLATE_DIR = Path(__file__).parent.parent / "templates"
-jinja_env = jinja2.Environment(
-    loader=jinja2.FileSystemLoader(str(TEMPLATE_DIR)),
-    autoescape=jinja2.select_autoescape(),
-    trim_blocks=True,
-    lstrip_blocks=True,
-)
-
-
-def handle_mcp_exception(e: Exception, operation: str) -> HTTPException:
-    """
-    Map MCP exceptions to HTTP responses using their status codes.
-
-    Args:
-        e: The caught exception
-        operation: Description of operation (e.g., "search", "extract", "query")
-
-    Returns:
-        HTTPException with appropriate status code and message
-    """
-    if isinstance(e, MCPError):
-        # All MCP errors have status_code set by their __init__
-        status_code = e.status_code if e.status_code else 500
-
-        # For MCPApplicationError, use the detail attribute directly to avoid redundant prefixes
-        # For other errors, use the string representation
-        if isinstance(e, MCPApplicationError):
-            message = e.detail
-        else:
-            message = str(e)
-
-        # Map MCP authentication errors (401, 403) to 502 Bad Gateway
-        # These are external service auth issues, not user session issues
-        # This prevents the frontend from logging users out when MCP tools have auth problems
-        if status_code in {401, 403} and e.category == "application":
-            status_code = 502
-            message = f"MCP tool authentication failed: {message}"
-
-        # Log based on severity (client errors vs server errors)
-        original_error_name = type(e.original_error).__name__ if e.original_error else None
-        if status_code >= 500:
-            logger.error(
-                f"MCP {operation} error [{e.category}] ({status_code}): {message}",
-                exc_info=True,
-                extra={"category": e.category, "original_error": original_error_name},
-            )
-        else:
-            logger.warning(
-                f"MCP {operation} error [{e.category}] ({status_code}): {message}",
-                extra={"category": e.category, "original_error": original_error_name},
-            )
-
-        return HTTPException(status_code=status_code, detail=message)
-
-    # Non-MCP errors
-    logger.error(f"Unexpected error in MCP {operation}: {str(e)}", exc_info=True)
-    return HTTPException(
-        status_code=500,
-        detail=f"An unexpected error occurred during {operation}. Please try again.",
-    )
-
-
-def _get_agent_class():
-    """
-    Determine which agent class to use based on RhesisClient availability.
-
-    Returns:
-        ObservableMCPAgent if RhesisClient is available and not disabled, otherwise MCPAgent
-    """
-    from rhesis.backend.app.utils.observability import rhesis_client
-
-    if rhesis_client is not None and not getattr(rhesis_client, "is_disabled", False):
-        logger.info("Using ObservableMCPAgent for MCP operations (observability enabled)")
-        return ObservableMCPAgent
-    else:
-        logger.info("Using standard MCPAgent for MCP operations (observability disabled)")
-        return MCPAgent
-
-
-def _get_mcp_tool_config(db: Session, tool_id: str, organization_id: str, user_id: str = None):
-    """
-    Get MCP client and provider configuration from database by tool ID.
-
-    Args:
-        db: Database session
-        tool_id: Tool instance ID
-        organization_id: Organization ID (for authorization check)
-        user_id: User ID (for authorization check)
-
-    Returns:
-        Tuple of (MCPClient, provider_name, repository_context) ready to use.
-        repository_context is None or a dict with 'owner', 'repo', 'full_name'
-        for GitHub repository-scoped connections.
-
-    Raises:
-        MCPConfigurationError: If tool not found, deleted, not an MCP integration,
-            or invalid credentials
-    """
-    try:
-        tool = crud.get_tool(db, uuid.UUID(tool_id), organization_id, user_id)
-    except ItemDeletedException:
-        raise MCPConfigurationError(
-            f"Tool '{tool_id}' has been deleted. Please re-import the source."
-        )
-
-    if not tool:
-        raise MCPConfigurationError(
-            f"Tool '{tool_id}' not found. Please add it in /integrations/tools"
-        )
-
-    # Verify tool type is MCP
-    if tool.tool_type.type_value != "mcp":
-        raise MCPConfigurationError(f"Tool '{tool.name}' is not an MCP integration")
-
-    # Get provider name for the client
-    provider = tool.tool_provider_type.type_value
-
-    # Parse credentials JSON
-    try:
-        credentials_dict = json.loads(tool.credentials)
-    except (json.JSONDecodeError, TypeError) as e:
-        raise MCPConfigurationError(f"Invalid credentials format for tool '{tool_id}': {e}")
-
-    # Extract repository context for GitHub provider
-    repository_context = None
-    if provider == "github" and tool.tool_metadata and "repository" in tool.tool_metadata:
-        repo_data = tool.tool_metadata["repository"]
-        if "owner" in repo_data and "repo" in repo_data:
-            repository_context = {
-                "owner": repo_data["owner"],
-                "repo": repo_data["repo"],
-                "full_name": repo_data.get(
-                    "full_name", f"{repo_data['owner']}/{repo_data['repo']}"
-                ),
-            }
-
-    # Check if tool uses custom provider (requires manual JSON config) or standard provider
-    if provider == "custom":
-        # Custom provider: requires tool_metadata with full JSON config
-        if not tool.tool_metadata:
-            raise MCPConfigurationError("Custom provider tools require tool_metadata configuration")
-
-        factory = MCPClientFactory.from_tool_config(
-            tool_config=tool.tool_metadata,
-            credentials=credentials_dict,
-        )
-    else:
-        # Standard provider: SDK constructs config from YAML templates
-        factory = MCPClientFactory.from_provider(
-            provider=provider,
-            credentials=credentials_dict,
-        )
-
-    client = factory.create_client(provider)
-    return client, provider, repository_context
-
-
-def _get_mcp_client_from_params(
-    provider_type_id: uuid.UUID,
-    credentials: Dict[str, str],
-    tool_metadata: Optional[Dict[str, Any]],
-    db: Session,
-    organization_id: str,
-    user_id: str = None,
-):
-    """
-    Get MCP client from parameters without requiring a tool in the database.
-
-    Args:
-        provider_type_id: UUID of the provider type (TypeLookup)
-        credentials: Dictionary of credential key-value pairs
-        tool_metadata: Optional tool metadata (required for custom providers)
-        db: Database session
-        organization_id: Organization ID (for authorization check)
-        user_id: User ID (for authorization check)
-
-    Returns:
-        MCPClient ready to use
-
-    Raises:
-        ValueError: If provider not found, invalid configuration, or missing required fields
-    """
-    # Fetch provider type from database
-    provider_type = crud.get_type_lookup(db, provider_type_id, organization_id, user_id)
-
-    if not provider_type:
-        raise ValueError(
-            f"Provider type '{provider_type_id}' not found. Please verify the provider_type_id."
-        )
-
-    # Get provider name for the client
-    provider = provider_type.type_value
-
-    # Check if provider uses custom provider (requires manual JSON config) or standard provider
-    if provider == "custom":
-        # Custom provider: requires tool_metadata with full JSON config
-        if not tool_metadata:
-            raise ValueError("Custom provider requires tool_metadata configuration")
-
-        factory = MCPClientFactory.from_tool_config(
-            tool_config=tool_metadata,
-            credentials=credentials,
-        )
-    else:
-        # Standard provider: SDK constructs config from YAML templates
-        factory = MCPClientFactory.from_provider(
-            provider=provider,
-            credentials=credentials,
-        )
-
-    client = factory.create_client(provider)
-    return client
-
-
-@endpoint(
-    name="search_mcp",
-    bind={
-        # DEVELOPMENT ONLY: get_test_context() provides test bindings for remote testing.
-        # Disabled by default in production (returns empty dict when env vars not set).
-        # Only used by Rhesis developers during development.
-        **get_test_context(),
-        "tool_id": os.getenv("RHESIS_TOOL_ID"),
-    },
-    request_mapping={
-        "query": "{{ input }}",
-    },
-    response_mapping={},
-)
-async def search_mcp(
-    query: str, tool_id: str, db: Session, organization_id: str, user_id: str
-) -> List[Dict[str, str]]:
-    """
-    Search MCP server for items matching query.
-
-    Uses an AI agent to intelligently search the MCP server and return
-    structured results. The agent automatically determines the best search
-    strategy based on the server's available tools.
-
-    Args:
-        query: Natural language search query (e.g., "Find pages about authentication")
-        tool_id: ID of the configured tool instance
-        db: Database session
-        organization_id: Organization ID for loading tools from database
-        user_id: User ID for retrieving default generation model
-
-    Returns:
-        List of dicts, each containing:
-        - id: Item identifier
-        - url: Direct link to the item
-        - title: Human-readable title
-
-    Raises:
-        ValueError: If search fails or returns invalid JSON
-
-    Example:
-        >>> results = await search_mcp(
-        ...     "pages created last week",
-        ...     "tool-uuid-123",
-        ...     db,
-        ...     org_id,
-        ...     user_id
-        ... )
-        >>> print(results[0]["title"])
-    """
-    if not user_id:
-        raise ValueError("user_id is required")
-
-    # Load MCP client from database tool configuration
-    client, provider, repository_context = _get_mcp_tool_config(
-        db, tool_id, organization_id, user_id
-    )
-
-    search_prompt = jinja_env.get_template("mcp_search_prompt.jinja2").render(
-        provider=provider, repository_context=repository_context
-    )
-
-    # Use dynamic agent class based on RhesisClient availability
-    AgentClass = _get_agent_class()
-    agent = AgentClass(
-        model=DEFAULT_GENERATION_MODEL,
-        mcp_client=client,
-        system_prompt=search_prompt,
-        max_iterations=10,
-        verbose=False,
-    )
-
-    result = await agent.run_async(query)
-
-    logger.info(f"Raw Agent output: {repr(result.final_answer)}")
-
-    # Parse response - agent returns list of items
-    # (Fatal errors raise exceptions before reaching here)
-    try:
-        parsed_results = json.loads(result.final_answer)
-        if not isinstance(parsed_results, list):
-            raise ValueError("Agent returned invalid format: expected a list of items")
-        return parsed_results
-    except json.JSONDecodeError as e:
-        raise ValueError(f"Agent returned invalid JSON: {str(e)}")
-
-
-@endpoint(
-    name="extract_mcp",
-    bind={
-        # DEVELOPMENT ONLY: get_test_context() provides test bindings for remote testing.
-        # Disabled by default in production (returns empty dict when env vars not set).
-        # Only used by Rhesis developers during development.
-        **get_test_context(),
-        "tool_id": os.getenv("RHESIS_TOOL_ID"),
-    },
-    request_mapping={
-        "item_url": "{{ input }}",  # Map required 'input' field to item_url
-        "item_id": "{{ item_id }}",  # Optional custom field for item_id
-    },
-    response_mapping={},
-)
-async def extract_mcp(
-    item_id: Optional[str] = None,
-    item_url: Optional[str] = None,
-    tool_id: str = None,
-    db: Session = None,
-    organization_id: str = None,
-    user_id: str = None,
-) -> str:
-    """
-    Extract full content from an MCP item as markdown.
-
-    Uses an AI agent to retrieve and convert item content to markdown format.
-    The agent navigates the item structure and extracts all relevant content
-    including text, headings, lists, and nested blocks.
-
-    Args:
-        item_id: Item identifier (optional, use if URL is not available)
-        item_url: Item URL (optional, preferred if available)
-        tool_id: ID of the configured tool instance
-        db: Database session
-        organization_id: Organization ID for loading tools from database
-        user_id: User ID for retrieving default generation model
-
-    Returns:
-        Full item content formatted as markdown string
-
-    Raises:
-        ValueError: If extraction fails, item not found, or neither id nor url provided
-
-    Example:
-        >>> content = await extract_mcp(
-        ...     item_id="page-id-123",
-        ...     tool_id="tool-uuid-123",
-        ...     db=db,
-        ...     organization_id=org_id,
-        ...     user_id=user_id
-        ... )
-        >>> print(content[:100])  # First 100 chars
-    """
-    if not item_id and not item_url:
-        raise ValueError("Either 'item_id' or 'item_url' must be provided")
-
-    if not user_id:
-        raise ValueError("user_id is required")
-
-    # Load MCP client and provider from database tool configuration
-    client, provider, _ = _get_mcp_tool_config(db, tool_id, organization_id, user_id)
-
-    extract_prompt = jinja_env.get_template("mcp_extract_prompt.jinja2").render(
-        item_id=item_id,
-        item_url=item_url,
-        provider=provider,
-    )
-
-    # Use dynamic agent class based on RhesisClient availability
-    AgentClass = _get_agent_class()
-    agent = AgentClass(
-        model=DEFAULT_GENERATION_MODEL,
-        mcp_client=client,
-        system_prompt=extract_prompt,
-        max_iterations=15,
-        verbose=False,
-    )
-
-    # Use URL if available, otherwise use ID for the prompt
-    item_reference = item_url if item_url else item_id
-    result = await agent.run_async(f"Extract content from item {item_reference}")
-
-    # Parse response - agent returns content as text
-    # (Fatal errors raise exceptions before reaching here)
-    return result.final_answer
-
-
-@endpoint(
-    name="query_mcp",
-    bind={
-        # DEVELOPMENT ONLY: get_test_context() provides test bindings for remote testing.
-        # Disabled by default in production (returns empty dict when env vars not set).
-        # Only used by Rhesis developers during development.
-        **get_test_context(),
-        "tool_id": os.getenv("RHESIS_TOOL_ID"),
-    },
-    request_mapping={
-        "query": "{{ input }}",
-    },
-    response_mapping={},
-)
-async def query_mcp(
-    query: str,
-    tool_id: str,
-    db: Session,
-    organization_id: str,
-    user_id: str,
-    system_prompt: Optional[str] = None,
-    max_iterations: int = 10,
-) -> Dict[str, Any]:
-    """
-    Execute arbitrary tasks on an MCP server using an AI agent.
-
-    Unlike search/extract endpoints, this provides full flexibility for any
-    MCP task with custom prompts and detailed execution traces. Use this for
-    complex operations like creating, updating, or analyzing content.
-
-    Args:
-        query: Natural language task description
-        tool_id: ID of the configured tool instance
-        db: Database session
-        organization_id: Organization ID for loading tools from database
-        user_id: User ID for retrieving default generation model
-        system_prompt: Custom agent instructions (optional)
-        max_iterations: Maximum reasoning steps (default: 10)
-
-    Returns:
-        Dict with execution details and result
-
-    Raises:
-        ValueError: If task execution fails
-
-    Example:
-        >>> result = await query_mcp(
-        ...     "Create a page titled 'Q1 Goals'",
-        ...     "tool-uuid-123", db, org_id, user_id
-        ... )
-    """
-    if not user_id:
-        raise ValueError("user_id is required")
-
-    # Load MCP client from database tool configuration
-    client, provider, repository_context = _get_mcp_tool_config(
-        db, tool_id, organization_id, user_id
-    )
-
-    if not system_prompt:
-        system_prompt = jinja_env.get_template("mcp_default_query_prompt.jinja2").render(
-            provider=provider, repository_context=repository_context
-        )
-
-    # Use dynamic agent class based on RhesisClient availability
-    AgentClass = _get_agent_class()
-    agent = AgentClass(
-        model=DEFAULT_GENERATION_MODEL,
-        mcp_client=client,
-        system_prompt=system_prompt,
-        max_iterations=max_iterations,
-        verbose=False,
-    )
-
-    result = await agent.run_async(query)
-
-    # Agent now raises exceptions for errors, so if we get here, it succeeded
-    return result.model_dump()
-
-
-async def run_mcp_authentication_test(
-    db: Session,
-    user: User,
-    organization_id: str,
-    tool_id: Optional[str] = None,
-    provider_type_id: Optional[uuid.UUID] = None,
-    credentials: Optional[Dict[str, str]] = None,
-    tool_metadata: Optional[Dict[str, Any]] = None,
-    user_id: str = None,
-) -> Dict[str, Any]:
-    """
-    Test MCP connection authentication by calling a tool that requires authentication.
-
-    Uses an AI agent to call a tool requiring authentication. The LLM determines
-    whether authentication is successful based on the tool call results.
-
-    Args:
-        db: Database session
-        user: Current user (for authorization)
-        organization_id: Organization ID for loading tools from database
-        tool_id: Optional ID of the configured tool instance (for existing tools)
-        provider_type_id: Optional UUID of the provider type (for non-existent tools)
-        credentials: Optional dictionary of credential key-value pairs (for non-existent tools)
-        tool_metadata: Optional tool metadata (for non-existent tools that use custom providers)
-        user_id: User ID for authorization check
-
-    Returns:
-        Dict with:
-        - is_authenticated: str - "Yes" or "No" determined by the LLM based on tool call results
-        - message: str - Message from the LLM explaining the authentication status
-        - additional_metadata: Optional[Dict[str, Any]] - Provider-specific metadata
-          (e.g., spaces for Jira)
-
-    Raises:
-        ValueError: If authentication test fails due to connection issues
-    """
-    # Load MCP client from either tool_id or parameters and get provider name
-    provider = None
-    if tool_id is not None:
-        client, provider, _ = _get_mcp_tool_config(db, tool_id, organization_id, user_id)
-    else:
-        # Get provider name from provider_type_id
-        provider_type = crud.get_type_lookup(db, provider_type_id, organization_id, user_id)
-        if provider_type:
-            provider = provider_type.type_value
-
-        client = _get_mcp_client_from_params(
-            provider_type_id=provider_type_id,
-            credentials=credentials,
-            tool_metadata=tool_metadata,
-            db=db,
-            organization_id=organization_id,
-            user_id=user_id,
-        )
-
-    # Load the authentication test prompt with provider context
-    auth_prompt = jinja_env.get_template("mcp_test_auth_prompt.jinja2").render(provider=provider)
-    agent = MCPAgent(
-        model=DEFAULT_GENERATION_MODEL,
-        mcp_client=client,
-        system_prompt=auth_prompt,
-        max_iterations=5,  # Keep it short for authentication test
-        verbose=False,
-    )
-
-    # Run agent with query to test authentication
-    query = "Call a tool that requires authentication to verify your credentials"
-    result = await agent.run_async(query)
-
-    if not result.success:
-        raise ValueError(f"Authentication test failed: {result.error}")
-
-    # Parse the response
-    response = json.loads(result.final_answer)
-
-    # Extract spaces if present (Jira-specific) and move to additional_metadata
-    additional_metadata = None
-    if "spaces" in response:
-        additional_metadata = {"spaces": response.pop("spaces")}
-
-    # Add additional_metadata to response
-    if additional_metadata:
-        response["additional_metadata"] = additional_metadata
-
-    return response
-
-
-async def create_jira_ticket_from_task(
-    task_id: uuid.UUID,
-    tool_id: str,
-    db: Session,
-    organization_id: str,
-    user_id: str,
-) -> Dict[str, Any]:
-    """
-    Create a Jira ticket from a task.
-
-    Args:
-        task_id: ID of the task to create a ticket from
-        tool_id: ID of the Jira MCP tool
-        db: Database session
-        organization_id: Organization ID
-        user_id: User ID
-
-    Returns:
-        Dict with issue_key, issue_url, and message
-
-    Raises:
-        ValueError: If task not found, tool not found, or tool not configured for Jira
-        MCPError: If ticket creation fails
-    """
-    # 1. Fetch task from database
-    task = crud.get_task(db, task_id, organization_id, user_id)
-    if not task:
-        raise ValueError(f"Task '{task_id}' not found")
-
-    # 2. Get MCP client and verify it's Jira
-    client, provider, _ = _get_mcp_tool_config(db, tool_id, organization_id, user_id)
-    if provider != "jira":
-        raise ValueError(f"Tool must be a Jira integration, got '{provider}'")
-
-    # 3. Extract space_key from tool metadata
-    tool = crud.get_tool(db, uuid.UUID(tool_id), organization_id, user_id)
-    if not tool.tool_metadata or "space_key" not in tool.tool_metadata:
-        raise ValueError("Jira tool is not configured with a space_key")
-
-    space_key = tool.tool_metadata["space_key"]
-
-    # 4. Prepare prompt with task data
-    create_issue_prompt = jinja_env.get_template("mcp_jira_create_issue_prompt.jinja2").render(
-        space_key=space_key,
-        summary=task.title,
-        description=task.description or "",
-    )
-
-    # 5. Use agent to create issue
-    agent = MCPAgent(
-        model=DEFAULT_GENERATION_MODEL,
-        mcp_client=client,
-        system_prompt=create_issue_prompt,
-        max_iterations=5,
-        verbose=False,
-    )
-
-    query = "Create the Jira issue as specified"
-    result = await agent.run_async(query)
-
-    if not result.success:
-        raise ValueError(f"Failed to create Jira ticket: {result.error}")
-
-    # 6. Parse response
-    response_data = json.loads(result.final_answer)
-
-    # 7. Update task_metadata with Jira issue information
-    if not task.task_metadata:
-        task.task_metadata = {}
-
-    task.task_metadata["jira_issue"] = {
-        "issue_key": response_data["issue_key"],
-        "issue_url": response_data["issue_url"],
-        "tool_id": tool_id,
-        "created_at": datetime.utcnow().isoformat(),
-    }
-
-    # Save updated task
-    crud.update_task(
-        db=db,
-        task_id=task_id,
-        task=schemas.TaskUpdate(task_metadata=task.task_metadata),
-        organization_id=organization_id,
-        user_id=user_id,
-    )
-
-    # 8. Return response
-    return response_data
